requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
Flask_PyMongo==2.3.0
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.3
